refactor: phase 3 bundle B — dedup post URL, OG helpers, cache post reads
B2: extract postUrl(slug) helper, replace 4 absolute-URL sites
B3: hoist identical loadProfile/loadFont into lib/og.ts
B4: wrap getAllPosts/getPostBySlug in React cache() for request-level dedup
