+++ app/opengraph-image.tsx
@@ -1,22 +1,10 @@
-import fs from "node:fs";
-import path from "node:path";
 import { ImageResponse } from "next/og";
 import { siteConfig } from "@/lib/site-config";
+import { loadProfile, loadFont } from "@/lib/og";
 
 export const size = { width: 1200, height: 630 };
 export const contentType = "image/png";
 export const alt = `${siteConfig.alternateName}`;
-
-function loadProfile() {
-  const file = path.join(process.cwd(), "public", "images", "profile.png");
-  const buffer = fs.readFileSync(file);
-  return `data:image/png;base64,${buffer.toString("base64")}`;
-}
-
-function loadFont() {
-  const file = path.join(process.cwd(), "public", "fonts", "Pretendard-Bold.otf");
-  return fs.readFileSync(file);
-}
 
 export default async function OpengraphImage() {
   const profile = loadProfile();
