+++ app/posts/[slug]/opengraph-image.tsx
@@ -1,25 +1,13 @@
-import fs from "node:fs";
-import path from "node:path";
 import { ImageResponse } from "next/og";
 import { notFound } from "next/navigation";
 import { getPostBySlug } from "@/lib/markdown";
 import { formatDate } from "@/lib/utils";
 import { siteConfig } from "@/lib/site-config";
+import { loadProfile, loadFont } from "@/lib/og";
 
 export const alt = "Post cover";
 export const size = { width: 1200, height: 630 };
 export const contentType = "image/png";
-
-function loadProfile() {
-  const file = path.join(process.cwd(), "public", "images", "profile.png");
-  const buffer = fs.readFileSync(file);
-  return `data:image/png;base64,${buffer.toString("base64")}`;
-}
-
-function loadFont() {
-  const file = path.join(process.cwd(), "public", "fonts", "Pretendard-Bold.otf");
-  return fs.readFileSync(file);
-}
 
 export default async function PostOpengraphImage({
   params,
