+++ app/sitemap.ts
@@ -1,6 +1,6 @@
 import type { MetadataRoute } from "next";
 import { getAllPosts } from "@/lib/markdown";
-import { siteConfig } from "@/lib/site-config";
+import { siteConfig, postUrl } from "@/lib/site-config";
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   const posts = (await getAllPosts()).filter((p) => !p.external);
@@ -20,7 +20,7 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       priority: 0.8,
     },
     ...posts.map((post) => ({
-      url: `${siteConfig.url}/posts/${post.slug}`,
+      url: postUrl(post.slug),
       lastModified: new Date(post.date),
       changeFrequency: "monthly" as const,
       priority: 0.7,
